2_cnn_fais_soln/app/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
faiss-cpu
numpy
